requirements.txt
streamlit
langchain-community
langchain-core
langchain-huggingface
langchain-groq
python-dotenv
beautifulsoup4
faiss-cpu
requests
smtplib